case-studies/clawdbot-campaign/campaign_orchestrator.py
#!/usr/bin/env python3
# Rights Reserved, Unlicensed
# mental-health-on-chain Campaign Orchestrator
# Autonomous marketing automation via clawdbot

import os
import json
import logging
import time
from datetime import datetime
from dotenv import load_dotenv

load_dotenv()

logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s [%(levelname)s] %(message)s',
    handlers=[
        logging.FileHandler('/opt/clawdbot/logs/campaign.log'),
        logging.StreamHandler()
    ]
)

logger = logging.getLogger(__name__)

# Load environment variables
REPO = "mental-health-on-chain"
GITHUB_ORG = "Future-Systems-Lab"
CONTRACT_ADDRESS = "0xE092336F8f5082e57CcBb341A110C20ad186A324"
SEPOLIA_EXPLORER = f"https://testnet.routescan.io/address/{CONTRACT_ADDRESS}"

class CampaignOrchestrator:
    def __init__(self):
        self.repo = REPO
        self.org = GITHUB_ORG
        self.campaign_start = datetime.now()
        logger.info(f"Campaign Orchestrator initialized for {self.repo}")
    
    def log_status(self, message):
        logger.info(message)
        self.save_metrics()
    
    def save_metrics(self):
        metrics = {
            'campaign_start': self.campaign_start.isoformat(),
            'last_update': datetime.now().isoformat(),
            'channels': ['Reddit', 'Discord', 'Dev.to', 'Twitter', 'GitHub'],
            'status': 'active'
        }
        with open('/opt/clawdbot/logs/metrics.json', 'w') as f:
            json.dump(metrics, f, indent=2)
    
    def run(self):
        self.log_status(f"Starting mental-health-on-chain campaign")
        self.log_status(f"Repository: {self.org}/{self.repo}")
        self.log_status(f"Contract: {SEPOLIA_EXPLORER}")
        self.save_metrics()
        
        # Keep running
        while True:
            time.sleep(300)  # Update metrics every 5 minutes
            self.save_metrics()
            logger.info("Campaign active - metrics updated")

if __name__ == "__main__":
    orchestrator = CampaignOrchestrator()
    orchestrator.run()
