case-studies/clawdbot-campaign/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
tweepy==4.14.0
discord.py==2.3.2
streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0
